push

+++ config.sh
@@ -6,23 +6,22 @@ if [ ! -d $HOME/.config/kitty ]; then
 else
   echo "kitty config dir existed"
 fi
-ln -s ~/.dotfiles/tool/kitty/kitty.conf ~/.config/kitty/kitty.conf
+if [ -f $HOME/.config/kitty/kitty.conf ]
+then
+  ln -s ~/.dotfiles/tool/kitty/kitty.conf ~/.config/kitty/kitty.conf
+fi
 
-
-
-echo "config LS_COLORS"
-mkdir /tmp/LS_COLORS && curl -L https://api.github.com/repos/vpeopleonatank/LS_COLORS/tarball/master | tar xzf - --directory=/tmp/LS_COLORS --strip=1
-( cd /tmp/LS_COLORS && sh install.sh )
 
 echo "turn off tmux auto restore"
 touch ~/tmux_no_auto_restore
 
 # Symlink config
-echo "Symlink init.vim"
+echo "Symlink nvim config"
+if [ ! -d $HOME/.config/nvim ]; then
+  mkdir $HOME/.config/nvim
+fi
 sudo ln -s $HOME/.dotfiles/tool/vim/init.vim $HOME/.config/nvim/init.vim
-
 sudo ln -s $HOME/.dotfiles/tool/vim/coc-settings-nvim.json $HOME/.config/nvim/coc-settings.json
-echo "Symlink coc-settings-nvim.json done"
 sudo ln -s $HOME/.dotfiles/tool/vim/lua $HOME/.config/nvim/lua
 sudo ln -s $HOME/.dotfiles/tool/snippets $HOME/.config/snippets
 echo "Install neovim Plugins"
@@ -37,9 +36,10 @@ else
     echo "tpm is installed"
 fi
 echo 'source ~/.dotfiles/tool/tmux/config.tmux' >$HOME/.tmux.conf
-echo "Symlink Tmux configuration successfully ^~^"
 
-mkdir -p $HOME/.config/jesseduffield/lazygit
-sudo ln -s $HOME/.dotfiles/tool/lazygit/config.yml $HOME/.config/jesseduffield/lazygit/config.yml
-mkdir -p $HOME/.config/jesseduffield/lazygit
+echo "Symlink Tmux configuration successfully ^~^"
+if [ ! -d $HOME/.config/jesseduffield/lazygit ]; then
+then
+  mkdir -p $HOME/.config/jesseduffield/lazygit
+fi
 sudo ln -s $HOME/.dotfiles/tool/lazygit/config.yml $HOME/.config/jesseduffield/lazygit/config.yml
